requirements.txt
aiogram==3.0.0b3
telethon==1.25.4
python-dotenv==0.20.0
asyncpg==0.27.0